+++ source/Plugins/Imports/uPSI_API_Shell.pas
@@ -50,7 +50,7 @@ begin
   with CL.AddClassN(CL.FindClass('TAPIBase'), 'TAPI_Shell') do
   begin
     RegisterMethod
-      ('Function RunCaptureCommand( const aCommandLine : string; var aOutput : string) : Integer');
+      ('Function RunCommandCapture( const aCommandLine : string; var aOutput : string) : Integer');
     RegisterMethod
       ('Function RunCommandSilent( const aFilename : String; const aDirectory : string; const aParameters : String) : Integer');
     RegisterMethod
@@ -70,7 +70,7 @@ procedure RIRegister_TAPI_Shell(CL: TPSRuntimeClassImporter);
 begin
   with CL.Add(TAPI_Shell) do
   begin
-    RegisterMethod(@TAPI_Shell.RunCaptureCommand, 'RunCaptureCommand');
+    RegisterMethod(@TAPI_Shell.RunCommandCapture, 'RunCommandCapture');
     RegisterMethod(@TAPI_Shell.RunCommandSilent, 'RunCommandSilent');
     RegisterMethod(@TAPI_Shell.RunCommand, 'RunCommand');
   end;